Move layout to utils
