+++ src/magic_tiler/utils/layout.py
